requirements.txt
backports-abc==0.5
certifi==2018.11.29
chardet==3.0.4
dominate==2.3.5
futures==3.2.0
idna==2.8
interval==1.0.0
numpy==1.15.4
opencv-python==3.4.5.20
Pillow==5.4.1
python-dateutil==2.7.5
pyzmq==17.1.2
requests==2.21.0
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
SSIM-PIL==1.0.2
torch==1.0.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
urllib3==1.24.1
visdom==0.1.8.5
websocket-client==0.54.0
